Moved bankcontext to new folder.
