+++ BankBlazor.Api/Services/AccountService.cs
@@ -1,4 +1,5 @@
-﻿using BankBlazor.Api.Data.Entities;
+﻿using BankBlazor.Api.Data.Context;
+using BankBlazor.Api.Data.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Identity.Client;
 using System.Security.Principal;
