+++ BankBlazor.Api/Services/CustomerService.cs
@@ -1,4 +1,4 @@
-﻿using BankBlazor.Api.Data.Entities;
+﻿using BankBlazor.Api.Data.Context;
 using Microsoft.EntityFrameworkCore;
 
 namespace BankBlazor.Api.Services
